+++ application/src/main/java/com/ericsson/bss/cassandra/ecchronos/application/MeterFilterImpl.java
@@ -15,21 +15,25 @@
 
 package com.ericsson.bss.cassandra.ecchronos.application;
 
+import com.ericsson.bss.cassandra.ecchronos.application.config.Config;
 import io.micrometer.core.instrument.Meter;
+import io.micrometer.core.instrument.Tag;
 import io.micrometer.core.instrument.config.MeterFilter;
 import io.micrometer.core.instrument.config.MeterFilterReply;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.util.List;
+import java.util.Map;
 import java.util.Set;
 
 public class MeterFilterImpl implements MeterFilter
 {
     private static final Logger LOG = LoggerFactory.getLogger(MeterFilterImpl.class);
     private final String myPrefix;
-    private final Set<String> myExcludedMetrics;
+    private final Set<Config.ExcludedMetric> myExcludedMetrics;
 
-    public MeterFilterImpl(final String prefix, final Set<String> excludedMetrics)
+    public MeterFilterImpl(final String prefix, final Set<Config.ExcludedMetric> excludedMetrics)
     {
         myPrefix = prefix;
         myExcludedMetrics = excludedMetrics;
@@ -44,26 +48,70 @@ public class MeterFilterImpl implements MeterFilter
     @Override
     public MeterFilterReply accept(final Meter.Id id)
     {
-        String name = id.getName();
-        if (myExcludedMetrics == null || name == null)
+        String metricName = id.getName();
+        if (myExcludedMetrics == null || metricName == null)
         {
             return MeterFilterReply.NEUTRAL;
         }
-        if (myExcludedMetrics.contains(id.getName()))
+        metricName = removePrefixIfPresent(metricName);
+        List<Tag> tags = id.getTags();
+        for (Config.ExcludedMetric excludedMetric : myExcludedMetrics)
         {
-            LOG.trace("Metric with name '{}' matches excluded metrics, will filter this metric out.", name);
-            return MeterFilterReply.DENY;
-        }
-        for (String excludeMetricRegex : myExcludedMetrics)
-        {
-            if (name.matches(excludeMetricRegex))
+            if (shouldExclude(metricName, excludedMetric.getName(), tags, excludedMetric.getTags()))
             {
-                LOG.trace("Metric with name '{}' matches excluded metric regex '{}', will filter this metric out.",
-                        name, excludeMetricRegex);
                 return MeterFilterReply.DENY;
             }
         }
         return MeterFilterReply.NEUTRAL;
+    }
+
+    private String removePrefixIfPresent(final String metricName)
+    {
+        if (myPrefix != null && !myPrefix.isEmpty())
+        {
+            return metricName.replaceFirst(myPrefix + ".", "");
+        }
+        return metricName;
+    }
+
+    private boolean shouldExclude(final String metricName, final String excludeOnNameRegexp, final List<Tag> tags,
+            final Map<String, String> excludeOnTags)
+    {
+        if (metricName.matches(excludeOnNameRegexp))
+        {
+            // If no tags, then exclude on name
+            if (excludeOnTags == null || excludeOnTags.isEmpty())
+            {
+                LOG.trace("Excluding metric '{}' with tags '{}' it matches name regex '{}'",
+                        metricName, tags, excludeOnNameRegexp);
+                return true;
+            }
+            // If tags, then we must consider them together.
+            // I.e, metric matching only name but not tags should not be excluded.
+            if (shouldExcludeOnTags(tags, excludeOnTags))
+            {
+                LOG.trace("Excluding metric '{}' with tags '{}' it matches exclude [name regex '{}', tags '{}']",
+                        metricName, tags, excludeOnNameRegexp, excludeOnTags);
+                return true;
+            }
+        }
+        return false;
+    }
+
+    private boolean shouldExcludeOnTags(final List<Tag> tags, final Map<String, String> excludedTags)
+    {
+        int matches = 0;
+        for (Tag tag : tags)
+        {
+            for (Map.Entry<String, String> excludedTag : excludedTags.entrySet())
+            {
+                if (tag.getKey().equals(excludedTag.getKey()) && tag.getValue().matches(excludedTag.getValue()))
+                {
+                    matches++;
+                }
+            }
+        }
+        return matches >= excludedTags.size();
     }
 
     /**